+++ Server/TimeLineKeeping/src/main/java/com/timelinekeeping/model/Pair.java
@@ -39,4 +39,5 @@ public class Pair<K, V> {
                 ", value=" + value +
                 '}';
     }
+
 }