+++ Server/TimeLineKeeping/src/main/java/com/timelinekeeping/repository/NotificationRepo.java
@@ -1,12 +1,20 @@
 package com.timelinekeeping.repository;
 
 import com.timelinekeeping.entity.NotificationEntity;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 /**
  * Created by HienTQSE60896 on 9/20/2016.
  */
 @Repository
 public interface NotificationRepo extends JpaRepository<NotificationEntity, Long> {
+
+
+    public List<NotificationEntity> findByAccountReceive(Long accountid);
+
 }